+++ packages/core/lozenge/src/Lozenge/index.js
@@ -1,29 +1,26 @@
 // @flow
-import { Appearance } from '@atlaskit/theme';
+
+import { Theme } from '@atlaskit/theme';
 import React, { PureComponent, type Node } from 'react';
 import Container from './styledContainer';
 import Content from './styledContent';
-import { boldStyles, defaultStyles } from './theme';
-
-export type Appearances =
-  | 'default'
-  | 'success'
-  | 'removed'
-  | 'inprogress'
-  | 'new'
-  | 'moved'
-  | {};
+import { theme, type ThemeAppearance, type ThemeProps } from '../theme';
 
 type Props = {
-  /** Determines whether to apply the bold style or not. */
-  isBold?: boolean,
   /** The appearance type. */
-  appearance: Appearances,
-  /** max-width of lozenge container. Default to 200px. */
-  maxWidth?: number | string,
-  /** Elements to be rendered inside the lozenge. This should ideally be just
-   a word or two. */
+  appearance: ThemeAppearance,
+
+  /** Elements to be rendered inside the lozenge. This should ideally be just a word or two. */
   children?: Node,
+
+  /** Determines whether to apply the bold style or not. */
+  isBold: boolean,
+
+  /** max-width of lozenge container. Default to 200px. */
+  maxWidth: number | string,
+
+  /** The theme the component should use. */
+  theme: ThemeProps => ThemeProps,
 };
 
 export default class Lozenge extends PureComponent<Props> {
@@ -31,24 +28,22 @@ export default class Lozenge extends PureComponent<Props> {
     isBold: false,
     appearance: 'default',
     maxWidth: 200,
+    theme,
   };
 
   render() {
-    const { appearance, isBold, maxWidth, children } = this.props;
-
+    const { props } = this;
     return (
-      <Appearance
-        props={appearance}
-        theme={isBold ? boldStyles : defaultStyles}
-      >
-        {styleProps => (
-          <Container {...styleProps}>
-            <Content {...styleProps} maxWidth={maxWidth}>
-              {children}
-            </Content>
-          </Container>
-        )}
-      </Appearance>
+      <Theme values={props.theme}>
+        {({ lozenge }) => {
+          const themeProps = lozenge(props);
+          return (
+            <Container {...themeProps}>
+              <Content {...themeProps}>{props.children}</Content>
+            </Container>
+          );
+        }}
+      </Theme>
     );
   }
 }