+++ packages/core/navigation-next/src/renderer/__tests__/unit/components.js
@@ -2,7 +2,7 @@
 
 import React from 'react';
 import { mount, shallow } from 'enzyme';
-import ArrowRightIcon from '@atlaskit/icon/glyph/arrow-right';
+import ArrowRightCircleIcon from '@atlaskit/icon/glyph/arrow-right-circle';
 import Spinner from '@atlaskit/spinner';
 import { Provider } from 'unstated';
 
@@ -18,11 +18,13 @@ const mountWithProvider = element =>
 describe('navigation-next view renderer', () => {
   describe('Item', () => {
     it('should render the Item UI component', () => {
-      const wrapper = shallow(<Item id="id" />);
+      const wrapper = shallow(<Item text="Item" id="id" />);
       expect(wrapper.find(BaseItem)).toHaveLength(1);
     });
     it('should render a GoToItem if a goTo prop is passed', () => {
-      const withGoTo = mountWithProvider(<Item id="id" goTo="view" />);
+      const withGoTo = mountWithProvider(
+        <Item text="Item" id="id" goTo="view" />,
+      );
       expect(withGoTo.find(GoToItem)).toHaveLength(1);
     });
   });
@@ -32,12 +34,12 @@ describe('navigation-next view renderer', () => {
       const notInHoverState = mountWithProvider(
         <GoToItem id="id" goTo="view" />,
       );
-      expect(notInHoverState.find(ArrowRightIcon)).toHaveLength(0);
+      expect(notInHoverState.find(ArrowRightCircleIcon)).toHaveLength(0);
 
       const inHoverState = mountWithProvider(
         <GoToItem id="id" goTo="view" isHover />,
       );
-      expect(inHoverState.find(ArrowRightIcon)).toHaveLength(1);
+      expect(inHoverState.find(ArrowRightCircleIcon)).toHaveLength(1);
 
       // Confirm that the icon is being passed as the 'after' prop for the
       // underlying Item
@@ -47,7 +49,7 @@ describe('navigation-next view renderer', () => {
           .props()
           .after({ isActive: false, isHover: true, isSelected: false }),
       );
-      expect(itemAfter.find(ArrowRightIcon)).toHaveLength(1);
+      expect(itemAfter.find(ArrowRightCircleIcon)).toHaveLength(1);
     });
 
     it('should render a spinner in the after slot of the Item when the to prop matches the incomingView.id', () => {