+++ packages/growth/react-experiment-framework/src/ExperimentContext.js
@@ -1,0 +1,10 @@
+// @flow
+
+import { createContext } from 'react';
+import type { Experiments } from './types';
+
+const initialContext: Experiments = {};
+const Experiment = createContext(initialContext);
+
+export const ExperimentProvider = Experiment.Provider;
+export const ExperimentConsumer = Experiment.Consumer;