+++ packages/media/media-card/src/utils/analyticsUtils.ts
@@ -2,8 +2,8 @@ import { BaseAnalyticsContext } from '../index';
 import { version, name } from '../../package.json';
 
 export const getBaseAnalyticsContext = (
-  componentName,
-  actionSubjectId,
+  componentName: any,
+  actionSubjectId: any,
 ): BaseAnalyticsContext => ({
   packageVersion: version,
   packageName: name,